Class implementations have been sorted into different files, definitions have been corrected, and necessary overloads have been added.
